1.Disable debug mode 2.Remove binary files 3.Use deparments instead of querying for new search 4.Fix out_of_range exception when call stol for viewer count conversion

+++ include/youtube/api/channel.h
@@ -51,7 +51,7 @@ public:
 
     unsigned int video_count() const;
 
-    long view_count() const;
+    long long view_count() const;
 
     const std::string likes_playlist() const;
 
@@ -78,7 +78,7 @@ protected:
 
     unsigned int video_count_;
 
-    long         view_count_;
+    long long    view_count_;
 
     std::string likes_playlist_;
